third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-template-element/template-element/template-content-hierarcy-expected.txt

This is a testharness.js-based test.
[FAIL] Template content should throw exception when its ancestor in a different document but connected via host is being append.
  assert_equals: expected Document node with 0 children but got Document node with 2 children
Harness: the test ran to completion.
